refactor(api): refactor allProjects -> /projects
